Rename files from wgsl to tgpu
