+++ packages/typegpu/src/createRuntime.ts
@@ -8,6 +8,12 @@ import {
 } from './programBuilder';
 import type { TgpuSettable } from './settableTrait';
 import { TaskQueue } from './taskQueue';
+import {
+  type ExtractPlumValue,
+  type TgpuPlum,
+  type Unsubscribe,
+  isPlum,
+} from './tgpuPlum';
 import type {
   ComputePipelineExecutorOptions,
   ComputePipelineOptions,
@@ -15,20 +21,14 @@ import type {
   RenderPipelineOptions,
   SetPlumAction,
   TgpuRuntime,
-} from './typegpuRuntime';
-import { type AnyTgpuData, type TgpuAllocatable, isAllocatable } from './types';
-import {
-  type ExtractPlumValue,
-  type TgpuPlum,
-  type Unsubscribe,
-  isPlum,
-} from './wgslPlum';
-import type { TgpuSampler } from './wgslSampler';
+} from './tgpuRuntime';
+import type { TgpuSampler } from './tgpuSampler';
 import type {
   TgpuAnyTexture,
   TgpuAnyTextureView,
   TgpuTextureExternal,
-} from './wgslTexture';
+} from './tgpuTexture';
+import { type AnyTgpuData, type TgpuAllocatable, isAllocatable } from './types';
 
 /**
  * Holds all data that is necessary to facilitate CPU and GPU communication.
