+++ packages/typegpu/src/experimental/index.ts
@@ -9,10 +9,10 @@ export default tgpu;
 export * from '../errors';
 export * from '../types';
 export { AsCallable, Callable } from '../callable';
-export * from '../typegpuRuntime';
+export * from '../tgpuRuntime';
 export { default as ProgramBuilder, type Program } from '../programBuilder';
 export { StrictNameRegistry, RandomNameRegistry } from '../nameRegistry';
-export * from '../wgslBuiltin';
+export * from '../builtin';
 
 export { default as wgsl } from '../wgsl';
 export { std } from '../std';
@@ -21,29 +21,29 @@ export {
   isUsableAsStorage,
   isUsableAsUniform,
   isUsableAsVertex,
-} from '../wgslBuffer';
-export { asUniform, asReadonly, asMutable, asVertex } from '../wgslBufferUsage';
+} from '../tgpuBuffer';
+export { asUniform, asReadonly, asMutable, asVertex } from '../tgpuBufferUsage';
 
 export type {
   TgpuBuffer,
   Unmanaged,
-} from '../wgslBuffer';
+} from '../tgpuBuffer';
 export type {
   TgpuBufferUsage,
   TgpuBufferUniform,
   TgpuBufferReadonly,
   TgpuBufferMutable,
   TgpuBufferVertex,
-} from '../wgslBufferUsage';
-export type { TgpuCode } from '../wgslCode';
-export type { TgpuConst } from '../wgslConstant';
-export type { TgpuFn } from '../wgslFunction';
-export type { TgpuPlum } from '../wgslPlum';
+} from '../tgpuBufferUsage';
+export type { TgpuCode } from '../tgpuCode';
+export type { TgpuConst } from '../tgpuConstant';
+export type { TgpuFn } from '../tgpuFunction';
+export type { TgpuPlum } from '../tgpuPlum';
 export type { TgpuSettable } from '../settableTrait';
-export type { TgpuFn as TgpuFnExperimental } from '../wgslFunctionExperimental';
-export type { TgpuVar } from '../wgslVariable';
-export type { TgpuSampler } from '../wgslSampler';
+export type { TgpuFn as TgpuFnExperimental } from '../tgpuFunctionExperimental';
+export type { TgpuVar } from '../tgpuVariable';
+export type { TgpuSampler } from '../tgpuSampler';
 export type {
   TgpuTexture,
   TgpuTextureView,
-} from '../wgslTexture';
+} from '../tgpuTexture';
