+++ packages/typegpu/src/tgpuSlot.ts
